Added a test to check json<->yaml configs equality
Implemented a unit test that takes all the config pairs from doc/samples/tasks/
with identical names (e.g. nova/boot-and-delete.json and
nova/boot-and-delete.yaml) and checks that their contents are identical as
well. Also it tests that every config should have a pair (if there is *.json,
there also has to be *.yaml, and vice versa)

Change-Id: I4c0d185439e2a2c48a801fa8c62c7ed2e80a6031