Change install docs to fix wrong command
I fixed sample command for changing permission of directory which is exposed
to docker container.

Current docs command us to change permissions of "rally_db" directory.
But this directory is never created and used to being exposed.

So I changed it from "rally_db" to "/var/lib/rally_container" which
is exposed to container.

Change-Id: I61ffb4ce2c5546678bb97e2052bb196811b3e0bf